requirements.txt
anyio>=4.5
httpx>=0.27
httpx-sse>=0.4
pydantic>=2.7.2,<3.0.0
starlette>=0.27
python-multipart>=0.0.9
sse-starlette>=1.6.1
pydantic-settings>=2.5.2
uvicorn>=0.23.1
jsonschema>=4.20.0
typer>=0.12.4
python-dotenv>=1.0.0
rich>=13.9.4
websockets>=15.0.1
pyodbc>=5.0.0
sqlalchemy
fastmcp
mcp
fastapi